# api.yaml
openapi: 3.0.0
info:
  title: User API Documentation
  version: 1.0.0
  description: Express + MongoDB User CRUD APIs with Authentication

servers:
  - url: http://localhost:5000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      required:
        - name
        - email
        - password
        - phone
        - city
      properties:
        _id:
          type: string
          readOnly: true
          example: "67a3b2f1c2940f8d12a34ef0"
        name:
          type: string
          example: "Ritesh Vishwakarma"
        email:
          type: string
          example: "ritesh@example.com"
        password:
          type: string
          example: "P@ssw0rd"
        phone:
          type: string
          example: "9876543210"
        city:
          type: string
          example: "Pune"

    UserCreateRequest:
      type: object
      required:
        - name
        - email
        - password
        - phone
        - city
      properties:
        name:
          type: string
          example: "Ritesh Vishwakarma"
        email:
          type: string
          example: "ritesh@example.com"
        password:
          type: string
          example: "P@ssw0rd"
        phone:
          type: string
          example: "9876543210"
        city:
          type: string
          example: "Pune"

    UserResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/User"

    UsersListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"

    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Login successful"
        accessToken:
          type: string
          example: "your.jwt.token.here"

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "User not found"

paths:

  /login:
    post:
      summary: Login user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: "ritesh@example.com"
                password:
                  type: string
                  example: "P@ssw0rd"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Email not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    post:
      summary: Create new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: Get all users
      tags:
        - Users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"

  /users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: MongoDB ObjectId of user

    get:
      summary: Get user by ID
      tags:
        - Users
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Update user by ID
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete user by ID
      tags:
        - Users
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User deleted successfully"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"